New struct for mocking clients
